backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
neo4j==5.14.1
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-openai>=0.0.2

